Move date formatter to format
